+++ modules/android/ecommerce-app-1/app/src/main/java/com/ern/retailapp/view/adapter/SimilarProductsPagerAdapter.java
@@ -154,7 +154,7 @@ public class SimilarProductsPagerAdapter extends PagerAdapter {
 
         ((TextView) itemView.findViewById(R.id.category_discount))
                 .setText(Money.asIDR(
-                        BigDecimal.valueOf(Long.valueOf(CenterRepository
+                        BigDecimal.valueOf(Double.valueOf(CenterRepository
                                 .getCenterRepository().getMapOfProductsInCategory()
                                 .get(productCategory).get(position)
                                 .getSellMRP()))).toString());
